BACK-END STRUCTURE CHANGES + Models split in separate files for readability and scalability.
