+++ software-engineering/backend/inspektor/apps/core/models/__init__.py
@@ -1,0 +1,7 @@
+from .case import Case
+
+from .image import Image
+from .image import get_image_path  # TODO: Unnecessary exposure: is used in migration 0001_initial.py
+
+from .inspection import Inspection
+
